rename getUrl file for most read and radio schedules
