+++ src/app/routes/mostRead/getInitialData/index.js
@@ -1,5 +1,5 @@
 import fetchPageData from '../../utils/fetchPageData';
-import getMostReadEndpoint from '#lib/utilities/getMostReadUrls';
+import getMostReadEndpoint from '#lib/utilities/getMostReadUrl';
 
 export default async ({ service, variant }) => {
   const mostReadUrl = getMostReadEndpoint({ service, variant }).split('.')[0];
